Fix various warnings found by 'cl -c W4...':   misc.c(102,23): warning C4245: '=': conversion from 'int' to 'DWORD', signed/unsigned mismatch   pcrecv.c(257,13): warning C4245: '=': conversion from 'int' to 'unsigned int', signed/unsigned mismatch   swsvpkt.c(418,21): warning C4245: 'initializing': conversion from 'int' to 'DWORD', signed/unsigned mismatch   winmisc.c(199,32): warning C4245: '=': conversion from 'int' to 'ULONG_PTR', signed/unsigned mismatch   winpkt.c(120,8): warning C4245: 'initializing': conversion from 'int' to 'DWORD', signed/unsigned mismatch   get_ni.c(119,25): warning C4245: 'initializing': conversion from 'int' to 'sa_family_t', signed/unsigned mismatch   socket.c(638,14): warning C4245: 'return': conversion from 'int' to 'unsigned int', signed/unsigned mismatch   syslog.c(73,12): warning C4245: 'initializing': conversion from 'int' to 'unsigned long', signed/unsigned mismatch   syslog.c(100,12): warning C4245: 'initializing': conversion from 'int' to 'unsigned long', signed/unsigned mismatch   res_debu.c(283,14): warning C4456: declaration of 'n' hides previous local declaration   warning C4701: potentially uninitialized local variable 'expiry' used

+++ src/misc.c
@@ -98,8 +98,8 @@ int _w32_errno = 0;
   void cdecl CheckCpuType (void)
   {
     x86_type = 6;
-    x86_hard_math = 1;
-    x86_capability = -1;   /* All flags on! */
+    x86_hard_math  = 1;
+    x86_capability = (DWORD) -1;   /* All flags on! */
     strcpy (x86_vendor_id, "GenuineIntel");
   }
 #endif
